Updated configlet.py and as well merge db_updater.py into configlet.py.
